Improve ConfigurationManager
