+++ agent/agent-sentinel/pom.xml
@@ -13,14 +13,23 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.bithon.agent</groupId>
-      <artifactId>agent-controller</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.bithon.shaded</groupId>
       <artifactId>shaded-alibaba-sentinel</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
+
+    <!-- Controller Impl -->
+    <dependency>
+      <groupId>org.bithon.agent</groupId>
+      <artifactId>agent-controller</artifactId>
+    </dependency>
+
+    <!-- Metrics -->
+    <dependency>
+      <groupId>org.bithon.agent</groupId>
+      <artifactId>agent-observability</artifactId>
+    </dependency>
+
     <dependency>
       <groupId>org.bithon</groupId>
       <artifactId>component-brpc</artifactId>
@@ -45,7 +54,6 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>4.13.1</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
